Added unit test structure for os_xml
