+++ src/unit_tests/analysisd/test_decode-xml.c
@@ -18,56 +18,12 @@
 #include "../../analysisd/decoders/decode-xml.c"
 #include "../wrappers/common.h"
 
-const char * w_get_attr_val_by_name(xml_node * node, const char * name);
 bool w_get_attr_regex_type(xml_node * node, w_exp_type_t * type);
 int w_get_attr_offset(xml_node * node);
 
 /* setup/teardown */
 
 /* tests */
-
-// w_get_attr_val_by_name
-
-void w_get_attr_val_by_name_null_attr(void ** state) {
-
-    xml_node node = {0};
-    const char * retval;
-
-    retval = w_get_attr_val_by_name(&node, "test_attr_name");
-
-    assert_null(retval);
-}
-
-void w_get_attr_val_by_name_not_found(void ** state) {
-
-    xml_node node = {0};
-    char * attributes[] = {"attr_name_1", "attr_name_2", NULL};
-    char * values[] = {"attr_val_1", "attr_val_2", NULL};
-
-    node.attributes = attributes;
-    node.values = values;
-    const char * retval;
-
-    retval = w_get_attr_val_by_name(&node, "test_attr_name");
-
-    assert_null(retval);
-}
-
-void w_get_attr_val_by_name_found(void ** state) {
-
-    xml_node node = {0};
-    char * attributes[] = {"attr_name_1", "test_attr_name", "attr_name_3", NULL};
-    char * values[] = {"attr_val_1", "test_attr_value", "attr_val_3", NULL};
-
-    node.attributes = attributes;
-    node.values = values;
-    const char * retval;
-
-    retval = w_get_attr_val_by_name(&node, "test_attr_name");
-
-    assert_non_null(retval);
-    assert_string_equal(retval, "test_attr_value");
-}
 
 // w_get_attr_regex_type
 
@@ -239,12 +195,6 @@ void w_get_attr_offset_error(void ** state) {
 
 int main(void) {
     const struct CMUnitTest tests[] = {
-
-        // Test w_get_attr_val_by_name
-        cmocka_unit_test(w_get_attr_val_by_name_null_attr),
-        cmocka_unit_test(w_get_attr_val_by_name_not_found),
-        cmocka_unit_test(w_get_attr_val_by_name_found),
-
         // Test w_get_attr_regex_type
         cmocka_unit_test(w_get_attr_regex_type_not_found),
         cmocka_unit_test(w_get_attr_regex_type_osregex),
